refactor: split document list pane
